fix: dtype layer (float64→float32), extract _process_sequence, 48 new contract tests
Production changes:
- init_hTtime() returns float32 (was float64) — eliminates silent upcast overhead
- from_df() allocates volumes as float32 — halves memory for global volumes
- Extract _process_sequence() from train() — testable seam for the forward loop

Test coverage (289 → 337):
- test_feature_scaler.py: 18 tests (Green/Beige/Red) for all §3/§6 guarantees
- test_datasniffer_pure_state.py: 10 tests for sniff_pure_state_parity/schema
- test_config_validation.py: 20 tests for checksum laws, lifecycle, enum validation
- conftest.py: add missing HydraNetConfig fields, scope collection gate to full runs

CIC §10 updates: replace stale legacy_tests/ references in DataSniffer,
FeatureScaler, DataFetcher CICs with actual test file locations.

+++ tests/test_architecture.py
@@ -21,7 +21,7 @@ def test_architecture_hidden_state_init():
     h = model.init_hTtime(hidden_channels=32, H=H, W=W)
 
     assert h.shape == (1, 32, H, W)
-    assert h.dtype == torch.float64
+    assert h.dtype == torch.float32
     assert torch.all(h == 0)
